requirements.txt
flask==2.3.2
openai>=1.3.0
httpx>=0.24.1
python-dotenv==1.0.0
requests==2.31.0
